+++ tests/test_visitors/test_ast/test_keywords/test_loops/test_loop_with_else.py
@@ -3,7 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.violations.best_practices import (
-    RedundantForElseViolation,
+    RedundantLoopElseViolation,
 )
 from wemake_python_styleguide.visitors.ast.keywords import WrongLoopVisitor
 
@@ -49,6 +49,13 @@ else:
     ...
 """
 
+wrong_while_without_break = """
+while x > 2:
+    ...
+else:
+    ...
+"""
+
 # Correct:
 
 right_else_in_for_loop = """
@@ -90,7 +97,7 @@ else:
     ...
 """
 
-check_nested_if_else = """
+right_nested_if_else = """
 for x in '000':
     if x:
         ...
@@ -98,9 +105,16 @@ for x in '000':
         ...
 """
 
-while_with_break = """
+right_while_with_break = """
 while x > 2:
     break
+else:
+    ...
+"""
+
+right_while_without_break_and_else = """
+while x > 2:
+    ...
 """
 
 
@@ -110,6 +124,7 @@ while x > 2:
     wrong_nested_for_with_break,
     wrong_nested_while_with_break,
     wrong_multiple_breaks,
+    wrong_while_without_break,
 ])
 def test_wrong_else_in_for_loop(
     assert_errors,
@@ -123,7 +138,7 @@ def test_wrong_else_in_for_loop(
     visitor = WrongLoopVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [RedundantForElseViolation])
+    assert_errors(visitor, [RedundantLoopElseViolation])
 
 
 @pytest.mark.parametrize('code', [
@@ -131,8 +146,9 @@ def test_wrong_else_in_for_loop(
     right_nested_break_in_for_loop,
     right_multiple_nested_for_with_break,
     right_multiple_breaks,
-    check_nested_if_else,
-    while_with_break,
+    right_nested_if_else,
+    right_while_with_break,
+    right_while_without_break_and_else,
 ])
 def test_correct_else_in_for_loop(
     assert_errors,