+++ tests/test_visitors/test_filenames/test_module/test_module_name_length.py
@@ -3,6 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.violations.naming import (
+    TooLongNameViolation,
     TooShortNameViolation,
     WrongModuleNamePatternViolation,
 )
@@ -36,4 +37,32 @@ def test_length_option(assert_errors, assert_error_text, options):
     visitor.run()
 
     assert_errors(visitor, [TooShortNameViolation])
+    assert_error_text(visitor, filename.replace('.py', ''))
+
+
+@pytest.mark.parametrize('filename', [
+    'super_long_name_that_needs_to_be_much_shorter_to_fit_the_rule.py',
+    'package/another_ridiculously_lengthly_name_that_defies_this_rule.py',
+    '/root/please_do_not_ever_make_names_long_and_confusing_like_this.py',
+    'C:/hello_there_this_is_another_very_long_name_that_will_not_work.py',
+])
+def test_too_long_filename(assert_errors, filename, default_options):
+    """Testing that long file names are restricted."""
+    visitor = WrongModuleNameVisitor(default_options, filename=filename)
+    visitor.run()
+
+    assert_errors(visitor, [
+        TooLongNameViolation,
+    ])
+
+
+def test_max_length_option(assert_errors, assert_error_text, options):
+    """Ensures that option `--max-name-length` works."""
+    max_length = 55
+    filename = 'very_long_name_that_should_not_pass_unless_changed_shorter.py'
+    option_values = options(max_name_length=max_length)
+    visitor = WrongModuleNameVisitor(option_values, filename=filename)
+    visitor.run()
+
+    assert_errors(visitor, [TooLongNameViolation])
     assert_error_text(visitor, filename.replace('.py', ''))