+++ wemake_python_styleguide/options/defaults.py
@@ -21,6 +21,9 @@ from wemake_python_styleguide.types import Final
 #: Minimum variable's name length.
 MIN_NAME_LENGTH: Final = 2
 
+#: Maximum variable and module name length:
+MAX_NAME_LENGTH: Final = 55
+
 #: Whether you control ones who use your code.
 I_CONTROL_CODE: Final = True