+++ wemake_python_styleguide/visitors/ast/complexity/function.py
@@ -44,7 +44,7 @@ class _ComplexityCounter(object):
     def _update_variables(
         self,
         function: AnyFunctionDef,
-        variable: ast.Name,
+        variable_def: ast.Name,
     ) -> None:
         """
         Increases the counter of local variables.
@@ -53,15 +53,15 @@ class _ComplexityCounter(object):
         Check ``TooManyLocalsViolation`` documentation.
         """
         function_variables = self.variables[function]
-        if variable.id not in function_variables:
-            if variable.id == UNUSED_VARIABLE:
+        if variable_def.id not in function_variables:
+            if variable_def.id == UNUSED_VARIABLE:
                 return
 
-            parent = getattr(variable, 'parent', None)
+            parent = getattr(variable_def, 'parent', None)
             if isinstance(parent, self._not_contain_locals):
                 return
 
-            function_variables.append(variable.id)
+            function_variables.append(variable_def.id)
 
     def _check_sub_node(self, node: AnyFunctionDef, sub_node) -> None:
         is_variable = isinstance(sub_node, ast.Name)