+++ CHANGELOG.md
@@ -7,6 +7,7 @@ We used to have incremental versioning before `0.1.0`.
 
 ### Features
 
+- Adds `TooLongNameViolation`
 - **Breaking**: removes `--max-conditions` and `--max-elifs` options
 - **Breaking**: removes `--max-offset-blocks`
 - **Breaking**: changes default `TooManyConditionsViolation` threshold from `3` to `4`
@@ -14,6 +15,10 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to use `lambda` inside loops
 - Reserving names `self`, `cls`, and `mcs` for first arguments only
 - Forbids to use too many decorators
+- Now `RedundantLoopElseViolation` also checks `while` loops
+- Forbids to have unreachable code
+- Forbids to have statements that have no effect
+- Add `variable` to the blacklisted names
 
 ## Bugfixes